api/serializers1.py
# -*- coding:utf-8 -*-
from rest_framework import serializers
from .models1 import University, Student

class UniversitySerializer(serializers.ModelSerializer):
    class Meta:
        model = University
        fields = '__all__'



class StudentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Student
        fields = '__all__'